# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FAMILIES IN FILM APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8082/api/",
            "description": "LOCAL API Server"
        },
        {
            "url": "https://fifapiqa.essentialdemo.com/api",
            "description": "LIVE API Server"
        }
    ],
    "paths": {
...
        "/providerPaymentMethods": {
            "get": {
                "tags": [
                    "ProviderPaymentMethod"
                ],
                "summary": "Get a listing of the ProviderPaymentMethods.",
                "description": "Get all ProviderPaymentMethods",
                "operationId": "e6f44a7e3b15f323615c019a0d61497b",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "ProviderPaymentMethod"
                ],
                "summary": "Store a newly created ProviderPaymentMethod in storage",
                "description": "Store ProviderPaymentMethod",
                "operationId": "5fd156522e4a4751080ac37e3ccdf39f",
                "parameters": [
                    {
                        "name": "body",
                        "in": "path",
                        "description": "ProviderPaymentMethod that should be stored",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/providerPaymentMethods/{id}": {
            "get": {
                "tags": [
                    "ProviderPaymentMethod"
                ],
                "summary": "Display the specified ProviderPaymentMethod",
                "description": "Get ProviderPaymentMethod",
                "operationId": "ef8b341e0e9d4e242a21a99c6a054cba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ProviderPaymentMethod",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProviderPaymentMethod"
                ],
                "summary": "Update the specified ProviderPaymentMethod in storage",
                "description": "Update ProviderPaymentMethod",
                "operationId": "eceef89f03c8fe4a063f1512104073d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ProviderPaymentMethod",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "path",
                        "description": "ProviderPaymentMethod that should be updated",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProviderPaymentMethod"
                ],
                "summary": "Remove the specified ProviderPaymentMethod from storage",
                "description": "Delete ProviderPaymentMethod",
                "operationId": "a098b83291aa29d742730d4070a877b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of ProviderPaymentMethod",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}